Debate Github/requirements.txt
spacy
wikipedia
textblob
gTTS
playsound==1.2.2
SpeechRecognition
openai
google-generativeai
elevenlabs
azure-cognitiveservices-speech
simpleaudio
pydub
scikit-learn
gensim
matplotlib
seaborn
networkx
wordcloud
pygame
pyttsx3
beautifulsoup4